requirements.txt
openai
python-dotenv
pandas
numpy
matplotlib
pyyaml